rename python folder
