+++ .gitignore
@@ -1,7 +1,8 @@
-
+__pycache__
 unity/Assets/ML-Agents/Timers/
 unity\Assets\ML-Agents\Timers\
 python/Images/
 unity/Assets/ML-Agents/Timers/XLand_timers.json
 python/XLand/
 Datasets/
+*.out
