+++ main.py
@@ -1,14 +1,15 @@
 import numpy as np
 import time
 import gymnasium as gym
-import python
+import u3gym
 
 
 def time_f(n_envs):
     env = gym.vector.AsyncVectorEnv([
         lambda i=i: gym.make(
             "U3GymEnv-v0",
-            worker_id=i,
+            file_name='unity/Builds/LinuxTraining/XLand',
+            worker_id=i + 2,
             disable_env_checker=True,
             camera_width=64,
             camera_height=64
@@ -32,11 +33,7 @@ def time_f(n_envs):
 
 if __name__ == "__main__":
     ts = []
-    n_envs = [1, 4, 16, 64]
-    while len(ts) < len(n_envs):
-        try:
-            t = time_f(n_envs[len(ts)])
-            ts.append(t)
-            print(ts)
-        except Exception as e:
-            print("Failed for", n_envs[len(ts)], "trying again", e)
+    for n_envs in [1, 4, 16, 64]:
+        t = time_f(n_envs)
+        ts.append(t)
+        print(ts)
