--- python/depreciated/docker/push.sh
@@ -1,29 +1,0 @@
-#!/bin/bash
-# Copyright 2019 The SEED Authors
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-
-# Pushes docker image seed_rl:$LABEL to the URI $IMAGE_URI:$LABEL.
-
-set -e
-DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
-cd $DIR
-if test -z "$1"
-then
-  LABEL=$CONFIG
-else
-  LABEL=$1
-fi
-
-docker tag seed_rl:$LABEL $IMAGE_URI:$LABEL
-docker push $IMAGE_URI:$LABEL
