--- python/depreciated/environment_test.py
@@ -1,36 +1,0 @@
-# coding=utf-8
-# Copyright 2019 The SEED Authors
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-"""Environment test."""
-
-import os
-
-from absl import flags
-from seed_rl.atari import env
-import tensorflow as tf
-
-FLAGS = flags.FLAGS
-
-
-class EnvironmentTest(tf.test.TestCase):
-
-  def test_run_step(self):
-    environment = env.create_environment(0)
-    environment.reset()
-    environment.step(0)
-    environment.close()
-
-
-if __name__ == '__main__':
-  tf.test.main()
