--- python/depreciated/r2d2_main.py
@@ -1,67 +1,0 @@
-# coding=utf-8
-# Copyright 2019 The SEED Authors
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-
-"""R2D2 binary for Unity.
-
-Actor and learner are in the same binary so that all flags are shared.
-"""
-
-from absl import app
-from absl import flags
-from seed_rl.agents.r2d2 import learner
-from seed_rl.unity import env
-from seed_rl.unity import networks
-from seed_rl.common import actor
-from seed_rl.common import common_flags  
-import tensorflow as tf
-
-gpu_options = tf.compat.v1.GPUOptions(allow_growth=True)
-sess = tf.compat.v1.Session(config=tf.compat.v1.ConfigProto(gpu_options=gpu_options))
-tf.compat.v1.keras.backend.set_session(sess)
-
-FLAGS = flags.FLAGS
-
-# Optimizer settings.
-flags.DEFINE_float('learning_rate', 1e-5, 'Learning rate.')
-flags.DEFINE_float('adam_epsilon', 1e-3, 'Adam epsilon.')
-
-
-def create_agent(env_output_specs, num_actions):
-  return networks.DuelingLSTMDQNNet(
-      num_actions, env_output_specs.observation.shape)
-
-
-def create_optimizer(unused_final_iteration):
-  learning_rate_fn = lambda iteration: FLAGS.learning_rate
-  optimizer = tf.keras.optimizers.Adam(FLAGS.learning_rate,
-                                       epsilon=FLAGS.adam_epsilon)
-  return optimizer, learning_rate_fn
-
-
-def main(argv):
-  if len(argv) > 1:
-    raise app.UsageError('Too many command-line arguments.')
-  if FLAGS.run_mode == 'actor':
-    actor.actor_loop(env.create_environment)
-  elif FLAGS.run_mode == 'learner':
-    learner.learner_loop(env.create_environment,
-                         create_agent,
-                         create_optimizer)
-  else:
-    raise ValueError('Unsupported run mode {}'.format(FLAGS.run_mode))
-
-
-if __name__ == '__main__':
-  app.run(main)
