+++ setup.py
@@ -1,0 +1,5 @@
+"""Setups the project."""
+from setuptools import setup
+
+
+setup(name="u3gym", version="0.0.1a")
