+++ u3gym/__init__.py
@@ -3,5 +3,5 @@ import gymnasium as gym
 
 gym.register(
     "U3GymEnv-v0",
-    "python.gym_wrapper:U3GymEnv"
+    "u3gym.gym_wrapper:U3GymEnv"
 )
